feat: refactor organization of the app project
